ideater/requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
sqlalchemy>=2.0.9
alembic>=1.10.3
python-dotenv>=1.0.0
httpx>=0.24.0

# LLM and AI integration
openai>=0.27.4
langchain>=0.0.200
tiktoken>=0.3.3
sentence-transformers>=2.2.2
llama-index>=0.6.0

# Visualization tools
graphviz>=0.20.1
pydot>=1.4.2
matplotlib>=3.7.1
plotly>=5.14.1
mermaid-cli>=0.1.1
diagrams>=0.23.3

# Web UI
jinja2>=3.1.2
aiofiles>=23.1.0
python-multipart>=0.0.6
itsdangerous>=2.1.2

# Document generation
markdown>=3.4.3
pdfkit>=1.0.0
weasyprint>=59.0
mdx-truly-sane-lists>=1.3.0
pygments>=2.15.1

# GitHub integration
pygithub>=1.58.1
gitpython>=3.1.31

# Notion integration
notion-client>=2.0.0

# Multi-agent system
crewai>=0.1.0
agentops>=0.1.0

# Utilities
numpy>=1.24.3
pandas>=2.0.1
tqdm>=4.65.0
python-slugify>=8.0.1
pytz>=2023.3
pyyaml>=6.0
jsonschema>=4.17.3
rich>=13.3.5
typer>=0.9.0
click>=8.1.3

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
faker>=18.4.0

# Development tools
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.2.0
pre-commit>=3.3.1
psycopg2-binary>=2.9.9
